requirements.txt
APScheduler==3.6.3
attrs==20.3.0
certifi==2020.6.20
chardet==3.0.4
decorator==4.0.11
filetype==1.0.5
future==0.18.2
huepy==1.2.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
iniconfig==1.1.1
instabot==0.117.0
instagram-private-api==1.6.0.0
instagram-private-api-extensions==0.3.8
instapy-cli==0.0.12
mock==4.0.2
moviepy==1.0.3
newsapi-python==0.2.6
numpy==1.19.3
packaging==20.4
Pillow==8.0.1
pluggy==0.13.1
proglog==0.1.9
py==1.9.0
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.1.2
pytz==2020.4
requests==2.24.0
requests-toolbelt==0.9.1
responses==0.12.1
schedule==0.6.0
selenium==3.141.0
six==1.15.0
toml==0.10.2
tqdm==4.51.0
urllib3==1.25.11
